+++ src/Commands.h
@@ -173,6 +173,8 @@ Cmd* enum (e.g. CmdOpen) and a human-readable name (not used yet).
     V(CmdPrevTab, "Previous Tab")                                                  \
     V(CmdNextTabSmart, "Smart Next Tab")                                           \
     V(CmdPrevTabSmart, "Smart Next Tab")                                           \
+    V(CmdMoveTabLeft, "Move Tab Left")                                             \
+    V(CmdMoveTabRight, "Move Tab Right")                                           \
     V(CmdSelectNextTheme, "Select next theme")                                     \
     V(CmdToggleFrequentlyRead, "Toggle Frequently Read")                           \
     V(CmdInvokeInverseSearch, "Invoke Inverse Search")                             \